refactor: move manage cursor logic to separate file
